@@ -11,22 +11,26 @@
         }
     ],
     "require": {
-        "twig/twig": "^1.27|^2.0"
+        "php": "^7.1.3",
+        "twig/twig": "^2.7",
+        "symfony/polyfill-mbstring": "^1.10"
     },
     "require-dev": {
-        "symfony/phpunit-bridge": "^3.4",
-        "symfony/translation": "^2.7|^3.4|^4.2"
+        "symfony/phpunit-bridge": "^4.2",
+        "symfony/translation": "^3.4|^4.2"
     },
     "suggest": {
         "symfony/translation": "Allow the time_diff output to be translated"
     },
     "autoload": {
-        "psr-0": { "Twig_Extensions_": "lib/" },
         "psr-4": { "Twig\\Extensions\\": "src/" }
+    },
+    "autoload-dev": {
+        "psr-4": { "Twig\\Tests\\Extensions\\": "test/" }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.5-dev"
+            "dev-master": "2.0-dev"
         }
     }
 }
