@@ -1,11 +1,156 @@
 <?php
 
+/*
+ * This file is part of Twig.
+ *
+ * (c) 2010-2019 Fabien Potencier
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Twig\Extensions\Node;
 
-class_exists('Twig_Extensions_Node_Trans');
+use Twig\Compiler;
+use Twig\Node\Expression\AbstractExpression;
+use Twig\Node\Expression\ConstantExpression;
+use Twig\Node\Expression\FilterExpression;
+use Twig\Node\Expression\NameExpression;
+use Twig\Node\Expression\TempNameExpression;
+use Twig\Node\Node;
+use Twig\Node\PrintNode;
 
-if (\false) {
-    class TransNode extends \Twig_Extensions_Node_Trans
+/**
+ * @author Fabien Potencier <fabien@symfony.com>
+ */
+class TransNode extends Node
+{
+    public function __construct(Node $body, Node $plural = null, AbstractExpression $count = null, Node $notes = null, $lineno, $tag = null)
     {
+        $nodes = ['body' => $body];
+        if (null !== $count) {
+            $nodes['count'] = $count;
+        }
+        if (null !== $plural) {
+            $nodes['plural'] = $plural;
+        }
+        if (null !== $notes) {
+            $nodes['notes'] = $notes;
+        }
+
+        parent::__construct($nodes, [], $lineno, $tag);
+    }
+
+    public function compile(Compiler $compiler)
+    {
+        $compiler->addDebugInfo($this);
+
+        list($msg, $vars) = $this->compileString($this->getNode('body'));
+
+        if ($this->hasNode('plural')) {
+            list($msg1, $vars1) = $this->compileString($this->getNode('plural'));
+
+            $vars = array_merge($vars, $vars1);
+        }
+
+        $function = $this->getTransFunction($this->hasNode('plural'));
+
+        if ($this->hasNode('notes')) {
+            $message = trim($this->getNode('notes')->getAttribute('data'));
+
+            // line breaks are not allowed cause we want a single line comment
+            $message = str_replace(["\n", "\r"], ' ', $message);
+            $compiler->write("// notes: {$message}\n");
+        }
+
+        if ($vars) {
+            $compiler
+                ->write('echo strtr('.$function.'(')
+                ->subcompile($msg)
+            ;
+
+            if ($this->hasNode('plural')) {
+                $compiler
+                    ->raw(', ')
+                    ->subcompile($msg1)
+                    ->raw(', abs(')
+                    ->subcompile($this->hasNode('count') ? $this->getNode('count') : null)
+                    ->raw(')')
+                ;
+            }
+
+            $compiler->raw('), array(');
+
+            foreach ($vars as $var) {
+                if ('count' === $var->getAttribute('name')) {
+                    $compiler
+                        ->string('%count%')
+                        ->raw(' => abs(')
+                        ->subcompile($this->hasNode('count') ? $this->getNode('count') : null)
+                        ->raw('), ')
+                    ;
+                } else {
+                    $compiler
+                        ->string('%'.$var->getAttribute('name').'%')
+                        ->raw(' => ')
+                        ->subcompile($var)
+                        ->raw(', ')
+                    ;
+                }
+            }
+
+            $compiler->raw("));\n");
+        } else {
+            $compiler
+                ->write('echo '.$function.'(')
+                ->subcompile($msg)
+            ;
+
+            if ($this->hasNode('plural')) {
+                $compiler
+                    ->raw(', ')
+                    ->subcompile($msg1)
+                    ->raw(', abs(')
+                    ->subcompile($this->hasNode('count') ? $this->getNode('count') : null)
+                    ->raw(')')
+                ;
+            }
+
+            $compiler->raw(");\n");
+        }
+    }
+
+    private function compileString(Node $body): array
+    {
+        if ($body instanceof NameExpression || $body instanceof ConstantExpression || $body instanceof TempNameExpression) {
+            return [$body, []];
+        }
+
+        $vars = [];
+        if (\count($body)) {
+            $msg = '';
+
+            foreach ($body as $node) {
+                if ($node instanceof PrintNode) {
+                    $n = $node->getNode('expr');
+                    while ($n instanceof FilterExpression) {
+                        $n = $n->getNode('node');
+                    }
+                    $msg .= sprintf('%%%s%%', $n->getAttribute('name'));
+                    $vars[] = new NameExpression($n->getAttribute('name'), $n->getTemplateLine());
+                } else {
+                    $msg .= $node->getAttribute('data');
+                }
+            }
+        } else {
+            $msg = $body->getAttribute('data');
+        }
+
+        return [new Node([new ConstantExpression(trim($msg), $body->getTemplateLine())]), $vars];
+    }
+
+    private function getTransFunction(bool $plural): string
+    {
+        return $plural ? 'ngettext' : 'gettext';
     }
 }
